# glotcast-api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/translate/transcript": {
      "post": {
        "operationId": "TranslateController_transcript_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateTranscriptDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptTranslationDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "translate"
        ]
      }
    },
    "/v1/translate/text": {
      "post": {
        "operationId": "TranslateController_text_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranslateTextDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TextTranslationDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "translate"
        ]
      }
    },
    "/v1/words/lookup": {
      "post": {
        "operationId": "WordsController_lookup_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LookupDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordLookupDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      }
    },
    "/v1/words": {
      "get": {
        "operationId": "WordsController_list_v1",
        "parameters": [
          {
            "name": "box",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 5,
              "type": "integer"
            }
          },
          {
            "name": "q",
            "required": false,
            "in": "query",
            "description": "matches the word or its meaning",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordsPageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      },
      "post": {
        "operationId": "WordsController_save_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SaveWordDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedWordDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      }
    },
    "/v1/words/stats": {
      "get": {
        "operationId": "WordsController_stats_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WordStatsDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      }
    },
    "/v1/words/review": {
      "get": {
        "operationId": "WordsController_due_v1",
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedWordsDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      }
    },
    "/v1/words/{id}": {
      "delete": {
        "operationId": "WordsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      }
    },
    "/v1/words/{id}/review": {
      "post": {
        "operationId": "WordsController_review_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReviewDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SavedWordDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "words"
        ]
      }
    },
...
  },
  "info": {
    "title": "GlotCast API",
    "description": "Podcasts in three English levels: catalog, listening, vocabulary, translation, and the admin content pipeline. The contract is docs/contract-v1.md; `admin`-tagged operations are for the admin panel only.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "Level_Output": {
        "type": "string",
        "enum": [
          "bg",
          "in",
          "ad"
        ],
        "description": "bg = beginner, in = intermediate, ad = advanced"
      },
...
      "Level": {
        "type": "string",
        "enum": [
          "bg",
          "in",
          "ad"
        ],
        "description": "bg = beginner, in = intermediate, ad = advanced"
      },
...
      "TranslateTranscriptDto": {
        "type": "object",
        "properties": {
          "episodeId": {
            "type": "string",
            "minLength": 1,
            "maxLength": 64,
            "pattern": "^[A-Za-z0-9-]+$"
          },
          "level": {
            "$ref": "#/components/schemas/Level"
          },
          "target": {
            "type": "string",
            "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
          }
        },
        "required": [
          "episodeId",
          "level",
          "target"
        ]
      },
      "TranscriptTranslationDto_Output": {
        "type": "object",
        "properties": {
          "target": {
            "type": "string"
          },
          "chunks": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "index-aligned with Transcript.chunks"
          }
        },
        "required": [
          "target",
          "chunks"
        ],
        "additionalProperties": false
      },
      "TranslateTextDto": {
        "type": "object",
        "properties": {
          "texts": {
            "minItems": 1,
            "maxItems": 100,
            "type": "array",
            "items": {
              "type": "string",
              "maxLength": 5000
            }
          },
          "target": {
            "type": "string",
            "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
          }
        },
        "required": [
          "texts",
          "target"
        ]
      },
      "TextTranslationDto_Output": {
        "type": "object",
        "properties": {
          "target": {
            "type": "string"
          },
          "texts": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        },
        "required": [
          "target",
          "texts"
        ],
        "additionalProperties": false
      },
      "LookupDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "target": {
            "type": "string",
            "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$",
            "description": "translation language"
          }
        },
        "required": [
          "word",
          "target"
        ]
      },
      "WordLookupDto_Output": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string"
          },
          "lemma": {
            "type": "string",
            "description": "the dictionary form the lookup used"
          },
          "phonetic": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "audioUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "translations": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "partOfSpeech": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "terms": {
                  "type": "array",
                  "items": {
                    "type": "string"
                  }
                }
              },
              "required": [
                "partOfSpeech",
                "terms"
              ],
              "additionalProperties": false
            }
          },
          "definitions": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "partOfSpeech": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "definition": {
                  "type": "string"
                },
                "example": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                }
              },
              "required": [
                "partOfSpeech",
                "definition",
                "example"
              ],
              "additionalProperties": false
            }
          }
        },
        "required": [
          "word",
          "lemma",
          "phonetic",
          "audioUrl",
          "translations",
          "definitions"
        ],
        "additionalProperties": false
      },
      "SavedWord_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "word": {
            "type": "string"
          },
          "meaning": {
            "type": "string"
          },
          "phonetic": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "audioUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "type": "string"
          },
          "box": {
            "type": "number",
            "enum": [
              1,
              2,
              3,
              4,
              5
            ]
          },
          "nextReviewAt": {
            "type": "string"
          },
          "lastReviewedAt": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "correctCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "incorrectCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WordSource_Output"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "word",
          "meaning",
          "phonetic",
          "audioUrl",
          "detail",
          "language",
          "box",
          "nextReviewAt",
          "lastReviewedAt",
          "correctCount",
          "incorrectCount",
          "createdAt",
          "source"
        ],
        "additionalProperties": false
      },
      "WordSource_Output": {
        "type": "object",
        "properties": {
          "episodeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "level": {
            "$ref": "#/components/schemas/Level_Output"
          },
          "chunkIndex": {
            "type": "integer",
            "minimum": 0,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "episodeId",
          "level",
          "chunkIndex"
        ],
        "additionalProperties": false
      },
      "WordsPageDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/SavedWord_Output"
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "WordStatsDto_Output": {
        "type": "object",
        "properties": {
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "byBox": {
            "type": "object",
            "properties": {
              "1": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "2": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "3": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "4": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "5": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              }
            },
            "required": [
              "1",
              "2",
              "3",
              "4",
              "5"
            ],
            "additionalProperties": false
          },
          "due": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          }
        },
        "required": [
          "total",
          "byBox",
          "due"
        ],
        "additionalProperties": false
      },
      "SavedWordsDto_Output": {
        "type": "array",
        "items": {
          "$ref": "#/components/schemas/SavedWord_Output"
        }
      },
      "SaveWordDto": {
        "type": "object",
        "properties": {
          "word": {
            "type": "string",
            "minLength": 1,
            "maxLength": 80
          },
          "meaning": {
            "type": "string",
            "maxLength": 500
          },
          "language": {
            "type": "string",
            "pattern": "^[A-Za-z]{2,3}(-[A-Za-z0-9]{2,8})*$"
          },
          "phonetic": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 120
              },
              {
                "type": "null"
              }
            ]
          },
          "audioUrl": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {},
          "source": {
            "anyOf": [
              {
                "type": "object",
                "properties": {
                  "episodeId": {
                    "type": "string",
                    "format": "uuid",
                    "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
                  },
                  "level": {
                    "$ref": "#/components/schemas/Level"
                  },
                  "chunkIndex": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  }
                },
                "required": [
                  "episodeId",
                  "level",
                  "chunkIndex"
                ]
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "word",
          "meaning",
          "language"
        ]
      },
      "SavedWordDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "word": {
            "type": "string"
          },
          "meaning": {
            "type": "string"
          },
          "phonetic": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "audioUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "detail": {
            "anyOf": [
              {},
              {
                "type": "null"
              }
            ]
          },
          "language": {
            "type": "string"
          },
          "box": {
            "type": "number",
            "enum": [
              1,
              2,
              3,
              4,
              5
            ]
          },
          "nextReviewAt": {
            "type": "string"
          },
          "lastReviewedAt": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "correctCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "incorrectCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "createdAt": {
            "type": "string"
          },
          "source": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/WordSource_Output"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "id",
          "word",
          "meaning",
          "phonetic",
          "audioUrl",
          "detail",
          "language",
          "box",
          "nextReviewAt",
          "lastReviewedAt",
          "correctCount",
          "incorrectCount",
          "createdAt",
          "source"
        ],
        "additionalProperties": false
      },
      "ReviewDto": {
        "type": "object",
        "properties": {
          "known": {
            "type": "boolean"
          }
        },
        "required": [
          "known"
        ]
      },
...
    }
  }
}